requirements.txt
fastapi
uvicorn[standard]
python-dotenv
anthropic
jinja2
python-multipart
